renamed com directory
