+++ app/src/androidTest/java/com/liaichi/gradeassignmenttracker/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.gymlog;
+package com.liaichi.gradeassignmenttracker;
 
 import android.content.Context;
 import androidx.test.platform.app.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
   public void useAppContext() {
     // Context of the app under test.
     Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-    assertEquals("com.example.gymlog", appContext.getPackageName());
+    assertEquals("com.example.gradeassignmenttracker", appContext.getPackageName());
   }
 }
