+++ app/src/main/java/com/liaichi/gradeassignmenttracker/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.gradeassignmenttracker;
+package com.liaichi.gradeassignmenttracker;
 
 import android.os.Bundle;
